requirements.txt
paho-mqtt
rrdtool
python-dateutil
astral
